update: cadastro de animal

See merge request evz-ufg-software-rural/bris-web!1

+++ bris-API/Models/Animal.cs
@@ -5,7 +5,7 @@ namespace bris_API.Models
         public int Id { get; set; }
         public string? Linhagem { get; set; }
         public int Idade { get; set; }
-        public int Peso { get; set; }
+        public float Peso { get; set; }
         public bool Status { get; set; }
         public int GranjaId { get; set; }
         public Granja? Granja { get; set; }
